jerry-bot/requirements.txt
backtrader>=1.9.78
alpaca-py>=0.37.0
polygon-api-client>=1.14.0
mplfinance>=0.12.10b0
tabulate>=0.9.0
matplotlib>=3.10.0
pandas>=2.2.3
numpy>=2.2.0
psutil>=6.1.1
